requirements.txt
gradio==5.33.0
torch==2.6.0
torchvision==0.21.0
pillow==11.1.0
pyttsx3==2.98
